Finalizando a entidade User

+++ JornadaMilhas.Application/Commands/DestinyCommands/RegisterDestiny/RegisterDestinyCommand.cs
@@ -10,5 +10,5 @@ using System.Threading.Tasks;
 
 namespace JornadaMilhas.Application.Commands.DestinyCommands.RegisterDestiny
 {
-    public sealed record RegisterDestinyCommand(string Name, string Subtitle, double Price, string DescriptionPortuguese, string DescriptionEnglish, List<string> Images) : IRequest<Result<Destiny>>{ }
+    public sealed record RegisterDestinyCommand(string Name, string Subtitle, decimal Price, string DescriptionPortuguese, string DescriptionEnglish, List<string> Images) : IRequest<Result<Destiny>>{ }
 }
